+++ HISTORY.rst
@@ -1,10 +1,42 @@
 Changelog
 ---------
 
+2.5.0
+~~~~~
+
+* Add cli option to specify html report directory
+
+* Fix use patch file under Windows
+
+* Improve documentation about cache issues
+
+2.4.5
+~~~~~
+
+* Add skipped column in html index report
+
+* Untested/skipped in show all does not show skipped
+
+* mutmut should pick up the config before running the original tests. That allows to specify test suits you want to run
+
+* improved some high-level help text & documentation
+
+2.4.4
+~~~~~
+
+* Fix issue with `TypeError` for empty file diffs
+
+* Introduce mutation of f-strings
+
+* Allow newer versions of junit-xml
+
 2.4.3
 ~~~~~
 
 * Fix for a big where worker cycling would just hang mutmut (thanks Jan Królikowski!)
+
+* Added `--CI` flag
+
 
 
 2.4.2
@@ -455,4 +487,3 @@ Thanks goes out Marcelo Da Cruz Pinto, Savo Kovačević,
 ~~~~~~~~~~~~~~~~~~
 
 * Initial version
-